refactor: rename and move modules and methods for better organization; add more docstrings.
