+++ nyx/moros_ecs/component/base_components.py
@@ -1,0 +1,9 @@
+from abc import ABC
+
+
+class BaseComponent(ABC):
+    """Abstract base class of all components in NyxEngine."""
+
+
+class RenderableComponent(BaseComponent, ABC):
+    """Abstract base class of all components that can be rendered by HemeraTerm"""
