--- nyx/moros_ecs/component/nyx_component.py
@@ -1,5 +1,0 @@
-from abc import ABC
-
-
-class NyxComponent(ABC):
-    """Abstract base class of all components in NyxEngine."""
