+++ nyx/moros_ecs/entity/nyx_entity.py
@@ -10,13 +10,20 @@ class NyxEntity:
         friendly_name (str, optional): The friendly name of this entity. Defaults to "".
     """
 
-    def __init__(self, friendly_name: int = ""):
-        """Initialize NyxEntity with an immutable globally unique entity id and optional friendly name"""
+    def __init__(self, friendly_name: str = ""):
+        """Initialize NyxEntity with an immutable globally unique entity id and optional friendly
+        name.
+
+        Args:
+            friendly_name (str, optional): The human-readable friendly name for the entity. Defaults
+                to "".
+        """
         self._entity_id: UUID = uuid4()
         self.friendly_name: str = friendly_name.strip() if friendly_name else ""
 
     @property
     def entity_id(self) -> UUID:
+        """Safely return the entity_id as a UUID."""
         return self._entity_id
 
     def __eq__(self, other):
@@ -29,8 +36,9 @@ class NyxEntity:
             return self.friendly_name == other or str(self.entity_id) == other
         return False
 
+    def __hash__(self):
+        """Use the entity_id when determining the hash of an entitiy."""
+        return hash(self.entity_id)
+
     def __repr__(self):
         return f"NyxEntity(friendly_name={self.name}, entity_id={self._entity_id})"
-
-    def __hash__(self):
-        return hash(self.entity_id)
